feat: refactor components, add paths for exports
