+++ package.json
@@ -13,7 +13,8 @@
     "@headlessui/react": "^1.7.14",
     "@heroicons/react": "^2.0.17",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "vite-tsconfig-paths": "^4.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
