--- src/componentExports.tsx
@@ -1,2 +1,0 @@
-export { default as Header } from "./components/Header";
-export { default as MobileMenu } from "./components/MobileMenu";
