+++ src/components/MobileMenu.tsx
@@ -15,15 +15,16 @@ const MobileMenu: React.FC<Props> = ({
 }) => {
   return (
     <Dialog
-      as="div"
+      as="nav"
       className="md:hidden"
       open={mobileMenuOpen}
       onClose={setMobileMenuOpen}
     >
       <div className="fixed inset-0 z-10" />
-      <Dialog.Panel className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-mongoose-300 px-6 py-6 sm:max-w-sm sm: shadow-lg">
+      <Dialog.Panel className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-mongoose-300 p-6">
         <div className="flex items-center justify-between">
-          <a href="#" className="-m-1.5 p-1.5">
+          <a href="#">
+            <h2 className="font-bold text-3xl">Gabriela Cresta</h2>
             {/* Waiting for logo */}
             {/* <img
                 className="h-8 w-auto"
@@ -33,23 +34,20 @@ const MobileMenu: React.FC<Props> = ({
           </a>
           <button
             type="button"
-            className="-m-2.5 rounded-md p-2.5 text-gray-900"
+            className="rounded-md text-gray-900"
             onClick={() => setMobileMenuOpen(false)}
           >
-            <XMarkIcon
-              className="h-6 w-6 hover:text-satin-linen"
-              aria-hidden="true"
-            />
+            <XMarkIcon className="h-6 w-6" aria-hidden="true" />
           </button>
         </div>
         <div className="mt-6 flow-root">
-          <div className="-my-6 divide-y divide-gray-500/10">
-            <div className="space-y-2 py-6">
+          <div className="divide-y divide-gray-500/10">
+            <div className="space-y-2">
               {navItems.map((item) => (
                 <a
                   key={item.name}
                   href={item.href}
-                  className="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-gray-900 hover:bg-satin-linen"
+                  className="block py-2 text-base font-semibold leading-7 text-gray-900"
                 >
                   {item.name}
                 </a>
