+++ src/components/index.tsx
@@ -1,0 +1,2 @@
+export { default as Header } from "@components/Header";
+export { default as MobileMenu } from "@components/MobileMenu";
